[#129] refactor : parseSentAt 유틸 함수로 분리
