+++ src/utils/date.ts
@@ -23,7 +23,7 @@ export function getTodayKstKey(): string {
   return kstIso.slice(0, 10); // "YYYY-MM-DD"
 }
 
-// UI 파싱용
+// UI 파싱용 (yyyy.mm.dd)
 export function getParseDate(iso: string) {
   const d = new Date(iso);
   const y = d.getFullYear();
@@ -31,3 +31,28 @@ export function getParseDate(iso: string) {
   const day = String(d.getDate()).padStart(2, '0');
   return `${y}.${m}.${day}`;
 }
+
+// UI 파싱용 (yyyy.mm.dd hh:mm am|pm)
+export const getParseSentAt = (isoOrNull: string | null) => {
+  if (!isoOrNull) return '-';
+
+  const d = new Date(isoOrNull);
+  if (Number.isNaN(d.getTime())) return '-';
+
+  const y = d.getFullYear();
+  const m = String(d.getMonth() + 1).padStart(2, '0');
+  const day = String(d.getDate()).padStart(2, '0');
+
+  let hours = d.getHours(); // 0 ~ 23
+  const minutes = String(d.getMinutes()).padStart(2, '0');
+
+  const isPM = hours >= 12;
+  const ampm = isPM ? 'PM' : 'AM';
+
+  hours = hours % 12;
+  if (hours === 0) hours = 12;
+
+  const hh = String(hours).padStart(2, '0');
+
+  return `${y}.${m}.${day} ${hh}:${minutes} ${ampm}`;
+};
